changed structure
I did not plan the file structure well. I may be lazy and have a dir of each type while core will be for the major game systems. I was thinking about merging controller with character, but controller may be used for other things so I am leaving it by itself.