update
Added a 2d version of the controller and character as well as a debug input (tab) to change between the 2d and 3d level. This is more for an example and testing different modes. 2d and 3d levels are different worlds (but could simulate the main level if needed). Also added a 3d tunnel that fake movement, but I might use a shader in the future and only have props that move if too complex for the shader.