update
Added a menu example and move gui to level instead of being an autoload. Most of the menu state is dependent on if the game is paused or not. A game state resource could be made if the main menu need to be independent of the pause state. Groups could work but the state be assumed. autoload and direct access be more direct, but that also means more dependent.

Also notice collision will be moved if setting its processing state to disable and then back to default state. There is a setting to treat it as a static instead of removed that is a fix for now but may be an issue of someone need to navigate (like a ghost) the world when paused. Truth is such case can be solved by using a different collision layer.