update
Added a physic state for keeping track of external influences of common physic variables. Also played with the idea of a movable body and decided to use character body as the base since static is too limited in movement features and rigid was too unreliable for a pushable box. This dynamic body would be a character body with some interactable functions and I may include a movement calculation resource later, but such object is not the goal of this project.